Made "show Scenario information" pass.
This has driven out the module, and author classes, at least in a
basic form.

I need to complete the relationship (i.e an author can have_many
scenarios), but for now, we are working.

Still not sure I've gone the right way with seasons, but again, for
now. It works.